pyproject.toml
[project]
name = "youtube-rag-backend"
version = "0.1.0"
description = "RAG backend for YouTube video chat app"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.120.1",
    "google>=3.0.0",
    "google-genai>=1.46.0",
    "httpx>=0.28.1",
    "langchain>=1.0.2",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
    "supabase>=2.22.2",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
    "youtube-transcript-api>=1.2.3",
]

[dependency-groups]
dev = [
    "poethepoet>=0.37.0",
]

[tool.poe.tasks]
hello = "echo Hello from poe!"
dev = "uvicorn app.main:app --port 8000 --reload"
prod = "uvicorn app.main:app"
